Fixed dialog windows tiling like main windows on tiling window managers.
I could not find in the documentation a list of what files are for dialog windows or anything along those lines, so I searched for instances of `self.setWindowModality` and as far as I'm seeing, it looks like that's all of them?
In any case, all I did was add `self.setWindowFlags(Qt.Dialog)` to the init functions of every modal I found that calls setWindowModality, the files affected are:
- add_field.py
- file_extension.py
- fix_dupes.py
- fix_unlinked.py
- folders_to_tags.py
- mirror_entities.py
- panel.py
- progress.py

Do let me know if I missed any or if some of these shouldn't be changed.

This fix has not been tested on Wayland yet, but based on the assumption that Qt behaves the same on both display servers, it should work the same, as this is a very basic feature that both display servers should have.

This fix is for [issue #392](https://github.com/TagStudioDev/TagStudio/issues/392)